# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.CreateSessionRequest:
    properties:
      department_id:
        type: string
      description:
        type: string
      end_time:
        type: string
      location:
        maxLength: 200
        type: string
      max_candidates:
        maximum: 500
        minimum: 1
        type: integer
      online_link:
        maxLength: 500
        type: string
      round:
        maximum: 20
        minimum: 1
        type: integer
      start_time:
        type: string
      title:
        maxLength: 200
        type: string
    required:
    - end_time
    - max_candidates
    - round
    - start_time
    - title
    type: object
...
  dto.DeleteResult:
    properties:
      deleted:
        type: integer
      keys:
        items:
          type: string
        type: array
    type: object
...
  dto.KeyInfo:
    properties:
      key:
        type: string
      ttl_seconds:
        type: integer
    type: object
...
  dto.SessionListResponse:
    properties:
      list:
        items:
          $ref: '#/definitions/dto.SessionView'
        type: array
      total:
        type: integer
    type: object
  dto.SessionView:
    properties:
      browser:
        type: string
      device:
        type: string
      expires_at:
        type: string
      ip:
        type: string
      login_at:
        type: string
      multi_device:
        type: boolean
      multi_ip:
        type: boolean
      os:
        type: string
      real_name:
        type: string
      token_id:
        type: string
      user_agent:
        type: string
      user_id:
        type: string
      username:
        type: string
    type: object
...
  dto.Stats:
    properties:
      healthy:
        type: boolean
      key_count:
        type: integer
      keys:
        items:
          $ref: '#/definitions/dto.KeyInfo'
        type: array
      l1_hits:
        type: integer
      l1_misses:
        type: integer
      l1_size:
        type: integer
      pattern:
        type: string
      pool:
        type: object
    type: object
...
  dto.UserSessionsResponse:
    properties:
      multi_device:
        type: boolean
      multi_ip:
        type: boolean
      real_name:
        type: string
      sessions:
        items:
          $ref: '#/definitions/dto.SessionView'
        type: array
      user_id:
        type: string
      username:
        type: string
    type: object
...
  dto.WarmupEntry:
    properties:
      key:
        type: string
      ttl_seconds:
        type: integer
      value:
        type: string
    required:
    - key
    type: object
  dto.WarmupRequest:
    properties:
      entries:
        items:
          $ref: '#/definitions/dto.WarmupEntry'
        type: array
      scan_prefix:
        type: string
    type: object
  dto.WarmupResult:
    properties:
      keys:
        items:
          type: string
        type: array
      loaded:
        type: integer
    type: object
...
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
      request_id:
        type: string
      timestamp:
        type: integer
    type: object
host: localhost:8080
info:
  contact: {}
  description: 高校计算机协会一体化管理平台
  title: StarByte API
  version: "1.0"
paths:
...
  /auth/sessions:
    get:
      description: 管理员查看当前在线 Access Token 会话，可按关键词或用户筛选
      parameters:
      - description: 关键词
        in: query
        name: keyword
        type: string
      - description: 用户 ID
        in: query
        name: user_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.SessionListResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 在线会话列表
      tags:
      - 认证
  /auth/sessions/{token}:
    delete:
      description: 按 token 标识强制下线一个会话
      parameters:
      - description: 会话 token 标识
        in: path
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 踢出单个会话
      tags:
      - 认证
  /auth/sessions/{user_id}:
    get:
      description: 查看某一用户当前全部在线会话
      parameters:
      - description: 用户 ID
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.UserSessionsResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 指定用户的在线会话
      tags:
      - 认证
  /auth/sessions/user/{user_id}:
    delete:
      description: 强制下线指定用户的全部在线会话
      parameters:
      - description: 用户 ID
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 踢出用户全部会话
      tags:
      - 认证
...
  /interviews/sessions:
    get:
      description: 场次列表
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 场次列表
      tags:
      - 面试
    post:
      consumes:
      - application/json
      description: 创建面试场次
      parameters:
      - description: 场次
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateSessionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 创建面试场次
      tags:
      - 面试
  /interviews/sessions/{id}:
    delete:
      description: 删除场次
      parameters:
      - description: 场次ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 删除场次
      tags:
      - 面试
    get:
      description: 场次详情
      parameters:
      - description: 场次ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 场次详情
      tags:
      - 面试
    put:
      consumes:
      - application/json
      description: 更新场次
      parameters:
      - description: 场次ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 更新场次
      tags:
      - 面试
...
  /system/cache/{key}:
    delete:
      description: 按完整键名删除 Redis 缓存
      parameters:
      - description: 缓存键
        in: path
        name: key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.DeleteResult'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 删除单个缓存键
      tags:
      - 缓存
  /system/cache/pattern/{pattern}:
    delete:
      description: 按 glob 模式批量删除缓存键
      parameters:
      - description: 匹配模式
        in: path
        name: pattern
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.DeleteResult'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 按模式清除缓存
      tags:
      - 缓存
  /system/cache/stats:
    get:
      description: 查看 Redis 连接池与按 pattern 扫描的键
      parameters:
      - description: 键匹配模式
        in: query
        name: pattern
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.Stats'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 缓存统计
      tags:
      - 缓存
  /system/cache/warmup:
    post:
      consumes:
      - application/json
      description: 将指定键写入 L1/L2 缓存
      parameters:
      - description: 预热条目
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.WarmupRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.WarmupResult'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 缓存预热
      tags:
      - 缓存
...
  /system/scheduler/handlers:
    get:
      description: 列出调度器已注册的 handler_key
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 可注册的任务处理器
      tags:
      - 调度
...
securityDefinitions:
  BearerAuth:
    description: 输入 Bearer {token}
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"